destroy session

+++ 0x03-user_authentication_service/app.py
@@ -48,14 +48,14 @@ def login():
 @app.route('/sessions', methods=['DELETE'], strict_slashes=False)
 def logout():
     """Logout route to destroy the session."""
-    session_cookie = request.cookies.get('session_id')
-    user = AUTH.get_user_from_session_id(session_cookie)
-
-    if user:
-        AUTH.destroy_session(user.id)
-        return redirect('/')
-    else:
-        abort(403)
+    if request.method == 'DELETE':
+        session_cookie = request.cookies.get('session_id')
+        user = AUTH.get_user_from_session_id(session_cookie)
+        if user:
+            AUTH.destroy_session(user.id)
+            return redirect('/')
+        else:
+            abort(403)
 
 
 if __name__ == "__main__":
